requirements.txt
Flask==1.1.2
SQLAlchemy==1.4.7
PyMySQL==1.0.2
gunicorn==20.1.0
requests==2.25.1
google-api-python-client==2.2.0
google-auth-oauthlib==0.4.4
google-auth==1.29.0
google-cloud-storage==1.37.1